+++ src/main/java/org/fatec/esportiva/mapper/ProductMapper.java
@@ -39,8 +39,8 @@ public class ProductMapper {
                 .build();
     }
 
-    public ProductResponseDto productResponseDto(Product product){
-        BigDecimal marginOfProfit = BigDecimal.ONE.add(BigDecimal.valueOf(product.getPricingGroup().getProfitMargin()));
+    public ProductResponseDto productResponseDto(Product product) {
+        BigDecimal marginOfProfit = BigDecimal.ONE.add(product.getPricingGroup().getProfitMargin());
         BigDecimal finalPrice = product.getCostValue().multiply(marginOfProfit).setScale(2, RoundingMode.HALF_UP);
         int availableQuantity = product.getStockQuantity() - product.getBlockedQuantity();
         return new ProductResponseDto(product.getId(), availableQuantity, product.getName(),
